+++ MediCore_API/Controllers/PatientController.cs
@@ -5,6 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using MediCore_API.Interfaces;
 using MediCore_API.Services;
 using MediCore_API.Models.DTOs;
+using Microsoft.AspNetCore.Identity;
 
 namespace MediCore_API.Controllers
 {
@@ -14,11 +15,13 @@ namespace MediCore_API.Controllers
     {
         private readonly MediCoreContext context;
         private readonly IModelMapper mapper;
+		private readonly UserManager<IdentityUser> userManager;
 
-        public PatientController(MediCoreContext context)
+		public PatientController(MediCoreContext context, UserManager<IdentityUser> userManager)
         {
             this.context = context;
             mapper = new ModelMapper();
+			this.userManager = userManager;
         }
 
         [HttpGet("/All")]
@@ -27,7 +30,6 @@ namespace MediCore_API.Controllers
 			try
 			{
 				var patients = await context.Patients.ToListAsync();
-				if (!patients.Any()) return NotFound("No Patients Found");
 				return Ok(patients.Select(p => mapper.Map<Patient, PatientDTO>(p)).ToList());
 			}
 			catch (Exception e)
@@ -51,12 +53,12 @@ namespace MediCore_API.Controllers
 			}
 		}
 
-        [HttpPatch("/{id:Guid}/Update")]
-        public async Task<ActionResult> PatchPatient([FromRoute] Guid id, [FromBody] PatientDTO dto)
+        [HttpPatch("/Update")]
+        public async Task<ActionResult> PatchPatient([FromBody] PatientDTO dto)
         {
 			try
 			{
-				var patient = await context.Patients.FirstOrDefaultAsync(p => p.Id == id);
+				var patient = await context.Patients.FirstOrDefaultAsync(p => p.Id == dto.Id);
 				if (patient is null) return NotFound("Patient Not Found");
 
 				if (!string.IsNullOrEmpty(dto.FirstName)) patient.FirstName = dto.FirstName;
@@ -82,7 +84,12 @@ namespace MediCore_API.Controllers
 			{
 				var patient = await context.Patients.FirstOrDefaultAsync(p => p.Id == id);
 				if (patient is null) return NotFound("Patient Not Found");
+				var user = await userManager.FindByIdAsync(patient.UserId);
+				if (user is null) return NotFound("User Not Found");
+
 				context.Patients.Remove(patient);
+				await userManager.DeleteAsync(user);
+
 				await context.SaveChangesAsync();
 				return Ok("Patient Deleted");
 			}
